Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)

Hello Chris,

The version numbers are maintained here:
https://gerrit.acumos.org/r/#/c/3101/1/modelbuilder/__init__.py

https://gerrit.acumos.org/r/gitweb?p=model-builder/h2o-model-builder.git;a=blob_plain;f=modelbuilder/_version.py;hb=HEAD
https://gerrit.acumos.org/r/#/c/3101/1/pom.xml
https://gerrit.acumos.org/r/gitweb?p=model-builder/h2o-model-builder.git;a=blob;f=setup.py;h=ae4cf702341a70e6a96f29078e42f2a2c4de3e3f;hb=HEAD

So if there is a better way to do this I'm open to whatever you suggest.   If you think we need to remove the comment in the pom file I'm fine with that too but I think it'd bite someone down the road not having it.

Patch-set: 1